app/services/pdf_analyzer.py
import pdfplumber
import fitz
import io
from PIL import Image
import pyzbar.pyzbar as pyzbar
from typing import Dict, Any, Tuple, List, Union
import logging

# Move PDFAnalyzer class here
class PDFAnalyzer:
    def generate_zpl_elements(self, dpi: int, width: float, height: float) -> Tuple[str, List[Dict[str, Any]]]:
        # ...existing code...

    def analyze_page(self, page_num: int = 0) -> Dict[str, Any]:
        # ...existing code...

    def _calculate_barcode_position(self, barcode: Any, image_pos: Dict[str, float], 
                                  image_size: Tuple[int, int]) -> Union[Dict[str, float], None]:
        # ...existing code...
