requirements.txt
fastapi
python-multipart
uvicorn
zebrafy==1.2.2
python-dotenv
pydantic<2.0.0
gunicorn
pillow>=9.1.0
weasyprint==59.0
cryptography
aiofiles
jinja2>=2.11.3
pdfplumber
requests
PyMuPDF
pyzbar
numpy
typing-extensions